Revert "Make sure images and horizontal rule are only added after the currently selected blcok"
This reverts commit dd8e34379f7cfb55e6e22dfd605d5d26be25e9e1.
